Move ~/src/midllewares/* to ~/src/middlewares/middlewares, #186
