+++ src/middlewares/middlewares/_.exporter.ts
@@ -1,0 +1,3 @@
+import { multerMiddlewareForComment, multerMiddlewareForProfile } from "./multer.middleware";
+
+export { multerMiddlewareForComment, multerMiddlewareForProfile };
