+++ src/routes/routers/profile.router.ts
@@ -2,7 +2,7 @@ import { Router } from "express";
 // import { applicationJsonFilter, formDataFilter } from "../../middlewares/filters/_.exporter";
 import { MulterProvider } from "../../modules/_.loader";
 import { preventUnLoginUserGuard } from "../../middlewares/guards/_.exporter";
-import { multerMiddlewareForProfile } from "../../middlewares/multer.middleware";
+import { multerMiddlewareForProfile } from "../../middlewares/middlewares/_.exporter";
 
 import ProfileController from "../controllers/profile.controller";
